feat: Sidebar api 연결
